Change the return type to Route
