+++ app/src/main/java/ru/hse/routemood/ServerApp.java
@@ -20,11 +20,11 @@ import ru.hse.routemood.gpt.TokenStore;
 @SpringBootApplication
 public class ServerApp {
     public static void main(String[] args) {
-        if (args.length < 1) {
-            throw new RuntimeException("No config file was given");
-        }
-        GptHandler.tokenFileName = args[0];
-        GptHandler.init();
+//        if (args.length < 1) {
+//            throw new RuntimeException("No config file was given");
+//        }
+//        GptHandler.tokenFileName = args[0];
+//        GptHandler.init();
         SpringApplication.run(ServerApp.class, args);
     }
 }
